+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add an SVG extension icon (#34)
 
 ### Changed
+- Use new instead of makeInstance for Tx_Oelib_List (#57)
 - Always use spaces for indentation (#43)
 - Require oelib >= 1.4.0 (#42)